[fs] nfs: Add method to detect whether an FSID is still on the server
Message-id: <1387216061-4641-32-git-send-email-steved@redhat.com>
Patchwork-id: 72347
O-Subject: [RHEL7.0 PATCH 31/53] NFS: Add method to detect whether an FSID is still on the server
Bugzilla: 769365
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Chuck Lever <chuck.lever@oracle.com>

Introduce a mechanism for probing a server to determine if an FSID
is present or absent.

The on-the-wire compound is different between minor version 0 and 1.
Minor version 0 appends a RENEW operation to identify which client
ID is probing.  Minor version 1 has a SEQUENCE operation in the
compound which effectively carries the same information.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Upstream Id: 44c9993384e9311cd56acf6ead3baffab616ae50
Bugzilla: 769365

Signed-off-by: Jarod Wilson <jarod@redhat.com>