[fs] nfs: Support NFS4ERR_LEASE_MOVED recovery in state manager
Message-id: <1387216061-4641-33-git-send-email-steved@redhat.com>
Patchwork-id: 72338
O-Subject: [RHEL7.0 PATCH 32/53] NFS: Support NFS4ERR_LEASE_MOVED recovery in state manager
Bugzilla: 769365
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Chuck Lever <chuck.lever@oracle.com>

A migration on the FSID in play for the current NFS operation
is reported via the error status code NFS4ERR_MOVED.

"Lease moved" means that a migration has occurred on some other
FSID than the one for the current operation.  It's a signal that
the client should take action immediately to handle a migration
that it may not have noticed otherwise.  This is so that the
client's lease does not expire unnoticed on the destination server.

In NFSv4.0, a moved lease is reported with the NFS4ERR_LEASE_MOVED
error status code.

To recover from NFS4ERR_LEASE_MOVED, check each FSID for that server
to see if it is still present.  Invoke nfs4_try_migration() if the
FSID is no longer present on the server.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Upstream Id: b7f7a66e420a9f3ec82a5124720013cee317e73a
Bugzilla: 769365

Signed-off-by: Jarod Wilson <jarod@redhat.com>