[fs] gfs2: Fix slab memory leak in gfs2_bufdata
Message-id: <a6858b3f6fed62a3de4350c9551bf54a148b26de.1387299605.git.rpeterso@redhat.com>
Patchwork-id: 72466
O-Subject: [RHEL7.0 PATCH 2/2] GFS2: Fix slab memory leak in gfs2_bufdata
Bugzilla: 1038123
RH-Acked-by: Benjamin Marzinski <bmarzins@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

This patch fixes a slab memory leak that sometimes can occur
for files with a very short lifespan. The problem occurs when
a dinode is deleted before it has gotten to the journal properly.
In the leak scenario, the bd object is pinned for journal
committment (queued to the metadata buffers queue: sd_log_le_buf)
but is subsequently unpinned and dequeued before it finds its way
to the ail or the revoke queue. In this rare circumstance, the bd
object needs to be freed from slab memory, or it is forgotten.
We have to be very careful how we do it, though, because
multiple processes can call gfs2_remove_from_journal. In order to
avoid double-frees, only the process that does the unpinning is
allowed to free the bd.

rhbz#1038123

Signed-off-by: Jarod Wilson <jarod@redhat.com>