[alsa] usb-audio: rename alt_idx to altsetting
Message-id: <138739275725-176-patchset-jkysela@redhat.com>
Patchwork-id: 72673
O-Subject: [RHEL 7.0 PATCH 175/365] ALSA: usb-audio: rename alt_idx to altsetting
Bugzilla: 1044022
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

As Clemens Ladisch kindly explained:
"Please note that there are two methods to identify alternate settings:
the number, which is the value in bAlternateSetting, and the index,
which is the index in the descriptor array.  There might be some wording
in the USB spec that these two values must be the same, but in reality,
[insert standard rant about firmware writers], bAlternateSetting
must be treated as a random ID value."

This patch changes the name to express the correct usage semantics.
No functional change.

Signed-off-by: Eldad Zack <eldad@fogrefinery.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: df23a2466a4961092625b487ead60f526e401062)
rhbz#1044022
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>