[alsa] usb: supply channel maps even when wChannelConfig is unspecified
Message-id: <138739275725-240-patchset-jkysela@redhat.com>
Patchwork-id: 72737
O-Subject: [RHEL 7.0 PATCH 239/365] ALSA: usb: supply channel maps even when wChannelConfig is unspecified
Bugzilla: 1044022
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

If wChannelconfig is given for some formats but not others, userspace
might not be able to set the channel map.

This is RFC because I'm not sure what the best behaviour is - to guess
the channel map from the given number of channels (it's quite likely
that one channel is MONO and two channels is FL FR), or just to supply
UNKNOWN for all channels.

But the complete lack of channel map for a format leads userspace to
believe that the format is not available at all. Or am I
misunderstanding how this should be used?

Signed-off-by: David Henningsson <david.henningsson@canonical.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 0dca01c37a68017fe7112b46bc4b48c927db18c6)
rhbz#1044022
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>