[alsa] usb: Don't trust the channel config if the channel count changed
Message-id: <138739275725-242-patchset-jkysela@redhat.com>
Patchwork-id: 72739
O-Subject: [RHEL 7.0 PATCH 241/365] ALSA: usb - Don't trust the channel config if the channel count changed
Bugzilla: 1044022
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

In case the channel count of the input terminal is not the same as
the channel count of the streaming descriptor, the channel config of
the input terminal can not be trusted. Instead fall back to a default
(guessed) channel map.

This was found on a Logitech USB Headset.

Signed-off-by: David Henningsson <david.henningsson@canonical.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 504333df8ba5fc310260285a22ab5d7cf3208795)
rhbz#1044022
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>