[alsa] hda: Enable mute/mic-mute LEDs for more Thinkpads with Realtek codec
Message-id: <138739275725-284-patchset-jkysela@redhat.com>
Patchwork-id: 72781
O-Subject: [RHEL 7.0 PATCH 283/365] ALSA: hda - Enable mute/mic-mute LEDs for more Thinkpads with Realtek codec
Bugzilla: 1044022
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

We're using the ACPI interface to detect whether we're dealing with a Thinkpad
or not. This way we're not loading the thinkpad_acpi module when we're not on
a Thinkpad, but at the same time, we give the opportunity to check for, and
potentially enable, both present and future Thinkpad with mute/micmute LEDs.

At least those running the ALC269 family (269 to 299) of Realtek codecs.

Cc: Alex Hung <alex.hung@canonical.com>
Tested-by: Hui Wang <hui.wang@canonical.com>
Signed-off-by: David Henningsson <david.henningsson@canonical.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 2793769f4450d9e4845257c953ae278ecdff7d94)
rhbz#1044022
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>