[redhat] kernel.spec: fix symlink dereference hardlinking
Apparently, cp -rl isn't legal, due to undefined behavior, and the latest
coreutils broke its usage entirely, but we can get back the functionality
we had by simply using cp -drl instead. This is a quick and easy fix to
just get people back to being able to build, we might actually look at
ripping out all the symlinking crud, since it was a Fedora thing that
isn't really so useful in RHEL, especially since we create a new tarball
for every build (it was a time-saver when recycling the same base tarball
for multiple builds, which doesn't apply here).

Signed-off-by: Jarod Wilson <jarod@redhat.com>