[ethernet] mlx4: Fix checking order in MR table init
Message-id: <94ae848326ef33bb51d35457e9505145c93d74d3.1388671195.git.avadai@redhat.com>
Patchwork-id: 73024
O-Subject: [Patch RHEL6 27/91] net/mlx4_core: Fix checking order in MR table init
Bugzilla: 1030563 1030565 1030568 1030570 1030571 1030573 1030575
RH-Acked-by: Doug Ledford <dledford@redhat.com>
RH-Acked-by: Jay Fenlason <fenlason@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jack Morgenstein <jackm@dev.mellanox.co.il>

In procedure mlx4_init_mr_table(), slaves should do no processing,
but should return success. This initialization is hypervisor-only.

However, the check for num_mpts being a power-of-2 was performed
before the check to return immediately if the driver is for a slave.
This resulted in spurious failures.

The order of performing the checks is reversed, so that if the
driver is for a slave, no processing is done and success is returned.

Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit a30f1bc5c0254cee94bc22304e57fe030fcf68f5)
Signed-off-by: Amir Vadai <amirv@mellanox.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>