[kernel] perf: Implement finer grained full dynticks kick
Message-id: <1389015820-2464-8-git-send-email-jeder@redhat.com>
Patchwork-id: 73225
O-Subject: [RHEL7.0 BZ1025775 v2 PATCH 07/13] perf: Implement finer grained full dynticks kick
Bugzilla: 1025775
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

From: Frederic Weisbecker <fweisbec@gmail.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1025775
Upstream: d84153d6c96f61aa06429586284639f32debf03e

Currently the full dynticks subsystem keep the
tick alive as long as there are perf events running.

This prevents the tick from being stopped as long as features
such that the lockup detectors are running. As a temporary fix,
the lockup detector is disabled by default when full dynticks
is built but this is not a long term viable solution.

To fix this, only keep the tick alive when an event configured
with a frequency rather than a period is running on the CPU,
or when an event throttles on the CPU.

These are the only purposes of the perf tick, especially now that
the rotation of flexible events is handled from a seperate hrtimer.
The tick can be shutdown the rest of the time.

Original-patch-by: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Frederic Weisbecker <fweisbec@gmail.com>
Cc: Jiri Olsa <jolsa@redhat.com>
Cc: Namhyung Kim <namhyung@kernel.org>
Cc: Arnaldo Carvalho de Melo <acme@redhat.com>
Cc: Stephane Eranian <eranian@google.com>
Signed-off-by: Peter Zijlstra <peterz@infradead.org>
Link: http://lkml.kernel.org/r/1374539466-4799-8-git-send-email-fweisbec@gmail.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>
(cherry picked from commit d84153d6c96f61aa06429586284639f32debf03e)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>