[tools] perf/gtk/hists: Set rules hint for the hist browser
Message-id: <1389109877-9395-49-git-send-email-jolsa@redhat.com>
Patchwork-id: 73347
O-Subject: [PATCH RHEL7.0 BZ1036665 048/226] perf gtk/hists: Set rules hint for the hist browser
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit 9d58d2f66c92fe47dd395947a3d51b9ace7dcc92
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Tue Jun 4 18:22:17 2013 +0900

description
===========
The 'rules' means that every second line of the tree view has a shaded
background, which makes it easier to see which cell belongs to which
row in the tree view.  It can be useful for a tree view that has a lot
of rows.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>