[tools] perf/symbols: Do not apply symfs for an absolute vmlinux path
Message-id: <1389109877-9395-67-git-send-email-jolsa@redhat.com>
Patchwork-id: 73368
O-Subject: [PATCH RHEL7.0 BZ1036665 066/226] perf symbols: Do not apply symfs for an absolute vmlinux path
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit 5698d2c99ec11006b4c241ebad5252289902b358
Author: Namhyung Kim <namhyung.kim@lge.com>
Date:   Wed Jul 17 17:08:15 2013 +0900

description
===========
If an user gives both of --symfs and --vmlinux option, the vmlinux will
be searched under the symfs directory.  This is somewhat confusing since
vmlinux often lives in kernel build directory or somewhere other than
user space binaries.

So it'd be better not adding symfs prefix for a vmlinux if it has an
absolute pathname.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>