[tools] perf: Support callchain sorting based on addresses
Message-id: <1389109877-9395-76-git-send-email-jolsa@redhat.com>
Patchwork-id: 73374
O-Subject: [PATCH RHEL7.0 BZ1036665 075/226] perf tools: Support callchain sorting based on addresses
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit 99571ab3d9b342a717295a9c7e2b4495ee19e32a
Author: Andi Kleen <ak@linux.intel.com>
Date: Thu Jul 18 15:33:57 2013 -0700

description
===========
With programs with very large functions it can be useful to distinguish
the callgraph nodes on more than just function names. So for example if
you have multiple calls to the same function, it ends up being separate
nodes in the chain.

This patch adds a new key field to the callgraph options, that allows
comparing nodes on functions (as today, default) and addresses.

Longer term it would be nice to also handle src lines, but that would
need more changes and address is a reasonable proxy for it today.

I right now reference the global params, as there was no simple way to
register a params pointer.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>