[tools] perf/annotate: Allow disassembly using /proc/kcore
Message-id: <1389109877-9395-107-git-send-email-jolsa@redhat.com>
Patchwork-id: 73407
O-Subject: [PATCH RHEL7.0 BZ1036665 106/226] perf annotate: Allow disassembly using /proc/kcore
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit bbb7f846f88df05646795854a014d73fb00f3b8b
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Wed Aug 7 14:38:54 2013 +0300

description
===========
Annotation with /proc/kcore is possible so the logic is adjusted to
allow it.  The main difference is that /proc/kcore had no symbols so the
parsing logic needed a tweak to read jump offsets.

The other difference is that objdump cannot always read from kcore.
That seems to be a bug with objdump.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>