[tools] perf/annotate: Improve description of '?' hotkey
Message-id: <1389109877-9395-111-git-send-email-jolsa@redhat.com>
Patchwork-id: 73410
O-Subject: [PATCH RHEL7.0 BZ1036665 110/226] perf annotate browser: Improve description of '?' hotkey
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit fcd9fef9a64b84520852e0247d6796893e28864c
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Aug 7 15:55:48 2013 -0300

description
===========
The previous description: "Search previous string" is usually associated
with the 'N' following a '/string', the opposite of 'n', which is
'Search next string' in the direction established with '/' or '?'.

So change it to 'Search string backwards', to clarify that.

The 'N' hotkey remains to be implemented with the semantic described
above.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>