[tools] perf/tests: Add tests of new pinned modifier
Message-id: <1389109877-9395-119-git-send-email-jolsa@redhat.com>
Patchwork-id: 73420
O-Subject: [PATCH RHEL7.0 BZ1036665 118/226] perf tests: Add tests of new pinned modifier
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit c9ee780f2736b7a149658b0cd8c8389da23e190a
Author: Michael Ellerman <michael@ellerman.id.au>
Date: Tue Aug 6 23:28:06 2013 +1000

description
===========
Add a negative test to test__checkevent_pmu_events() to get lots of
coverage of the negative case, ie. when the modifier is not specified.

Add a test of a single event, and of the group case.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>