[tools] perf/machine: Add symbol filter to struct machine
Message-id: <1389109877-9395-131-git-send-email-jolsa@redhat.com>
Patchwork-id: 73430
O-Subject: [PATCH RHEL7.0 BZ1036665 130/226] perf machine: Add symbol filter to struct machine
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit 611a5ce8aa4cdb2daefbd9bed77ec3b3e9bd00ea
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Aug 8 14:32:20 2013 +0300

description
===========
The symbol filter needs to be applied machine-wide, so add it to struct
machine.

Currently tools pass the symbol filter as a parameter to various
map-related functions.  However a need to load a map can occur anywhere
in the code, at which point the filter is needed.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>